% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_exp_quad_cpp}
\alias{cov_exp_quad_cpp}
\title{Compute an exponential quadratic covariance matrix, also known as a squared exponential}
\usage{
cov_exp_quad_cpp(xa, xb, sigma, l)
}
\arguments{
\item{xa}{\code{NumericVector} of values}

\item{xb}{\code{NumericVector} of values}

\item{sigma}{\code{double} denoting the variance}

\item{l}{\code{double} denoting the lengthscale}
}
\value{
\code{NumericMatrix} of covariance
}
\description{
Compute an exponential quadratic covariance matrix, also known as a squared exponential
}
\examples{
x1 <- seq(from = -2, to = 2, length.out = 100)
cov_exp_quad_cpp(x1, x1, 0.05, 1)
}
\author{
Trent Henderson
}
