% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_linear_cpp}
\alias{cov_linear_cpp}
\title{Compute a linear covariance matrix}
\usage{
cov_linear_cpp(xa, xb, sigma_b, sigma_v, c)
}
\arguments{
\item{xa}{\code{NumericVector} of values}

\item{xb}{\code{NumericVector} of values}

\item{sigma_b}{\code{double} denoting the constant variance}

\item{sigma_v}{\code{double} denoting the variance}

\item{c}{\code{double} denoting the offset}
}
\value{
\code{NumericMatrix} of covariance
}
\description{
Compute a linear covariance matrix
}
\examples{
x1 <- seq(from = -2, to = 2, length.out = 100)
cov_linear_cpp(x1, x1, 0.05, 1, 1)
}
\author{
Trent Henderson
}
