% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov_noise}
\alias{cov_noise}
\title{Compute a white noise covariance matrix}
\usage{
cov_noise(xa, xb, sigma = 0.5)
}
\arguments{
\item{xa}{\code{numeric} vector of values}

\item{xb}{\code{numeric} vector of values}

\item{sigma}{\code{numeric} scalar denoting the variance. Defaults to \code{0.5}}
}
\value{
\code{GPCov} containing the covariance matrix
}
\description{
Compute a white noise covariance matrix
}
\examples{
x1 <- seq(from = -2, to = 2, length.out = 100)
cov_noise(x1, x1, 0.5)

}
\author{
Trent Henderson
}
