% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_noise_cpp}
\alias{cov_noise_cpp}
\title{Compute a white noise covariance matrix}
\usage{
cov_noise_cpp(xa, xb, sigma)
}
\arguments{
\item{xa}{\code{NumericVector} of values}

\item{xb}{\code{NumericVector} of values}

\item{sigma}{\code{double} denoting the variance}
}
\value{
\code{NumericMatrix} of covariance
}
\description{
Compute a white noise covariance matrix
}
\examples{
x1 <- seq(from = -2, to = 2, length.out = 100)
cov_noise_cpp(x1, x1, 0.5)
}
\author{
Trent Henderson
}
