% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_rat_quad_cpp}
\alias{cov_rat_quad_cpp}
\title{Compute a rational quadratic covariance matrix}
\usage{
cov_rat_quad_cpp(xa, xb, sigma, alpha, l)
}
\arguments{
\item{xa}{\code{NumericVector} of values}

\item{xb}{\code{NumericVector} of values}

\item{sigma}{\code{double} denoting the variance}

\item{alpha}{\code{double} greater than \code{0} denoting the mixing coefficient}

\item{l}{\code{double} denoting the lengthscale}
}
\value{
\code{NumericMatrix} of covariance
}
\description{
Compute a rational quadratic covariance matrix
}
\examples{
x1 <- seq(from = -2, to = 2, length.out = 100)
cov_rat_quad_cpp(x1, x1, 0.05, 1, 1)
}
\author{
Trent Henderson
}
