% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_log_likelihood.R
\name{neg_log_likelihood}
\alias{neg_log_likelihood}
\title{Calculate negative marginal log likelihood of a Gaussian process regression model}
\usage{
neg_log_likelihood(params, x, xprime, y, covfun, epsilon = 1e-09)
}
\arguments{
\item{params}{\code{vector} of hyperparameters}

\item{x}{\code{numeric} vector of input data}

\item{xprime}{\code{numeric} vector of data points to predict}

\item{y}{\code{numeric} vector of values to learn from}

\item{covfun}{\code{function} specifying the covariance function to use}

\item{epsilon}{\code{numeric} scalar denoting a small quantity to add to the (x, x) covariance matrix for numerical stability. Defaults to \code{1e-9}}
}
\value{
\code{numeric} scalar of the negative log likelihood
}
\description{
Calculate negative marginal log likelihood of a Gaussian process regression model
}
\author{
Trent Henderson
}
