% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.GPCov}
\alias{plot.GPCov}
\title{Plot a covariance matrix}
\usage{
\method{plot}{GPCov}(x, xa, type = c("prior", "matrix"), k = 5, ...)
}
\arguments{
\item{x}{\code{GPCov} object containing the covariance matrix}

\item{xa}{\code{numeric} vector used to generate \code{x}}

\item{type}{\code{character} denoting whether to plot samples from the GP prior (\code{"prior"}) assuming a mean vector of zero, or to draw a heatmap of the covariance matrix (\code{"matrix"}). Defaults to \code{"prior"}}

\item{k}{\code{integer} denoting number of realisations to draw if \code{type = "prior"}. Defaults to \code{5}}

\item{...}{arguments to be passed to methods}
}
\value{
\code{ggplot2} object containing the plot
}
\description{
Plot a covariance matrix
}
\examples{
x1 <- 1:100
covmat <- cov_exp_quad(x1, x1, 0.05, 1)
plot(covmat, x1, type = "prior", k = 5)

}
\author{
Trent Henderson
}
