% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.TSGP}
\alias{plot.TSGP}
\title{Plot summary of draws from GP posterior}
\usage{
\method{plot}{TSGP}(x, prob = 0.95, draws = 100, ...)
}
\arguments{
\item{x}{\code{TSGP} object containing the model}

\item{prob}{\code{numeric} scalar denoting the probability for the credible interval. Defaults to \code{0.95} for 95% interval}

\item{draws}{\code{integer} denoting the number of draws to take from the posterior. Defaults to \code{100}}

\item{...}{arguments to be passed to methods}
}
\value{
\code{ggplot} object containing the plot
}
\description{
Plot summary of draws from GP posterior
}
\examples{
x1 <- 1:100
y <- 3 * sin(2 * seq(0, 4 * pi, length.out = 100)) + runif(100) * 2 + (0.08 * seq(from = 1, to = 100, by = 1))

CovSum <- function(xa, xb, sigma_1 = 1, sigma_2 = 1, l_1 = 1, l_2 = 1, p = 1){
  Sigma_exp_quad <- cov_exp_quad(xa, xb, sigma_1, l_1)
  Sigma_periodic <- cov_periodic(xa, xb, sigma_2, l_2, p)
  Sigma_noise <- cov_noise(sigma_3, nrow(Sigma_exp_quad))
  Sigma <- Sigma_exp_quad + Sigma_periodic
  return(Sigma)
}

mod <- GP(x1, 1:length(y), y, CovSum, 0.5)
plot(mod, 0.95, 100)

}
\author{
Trent Henderson
}
