% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov_noise}
\alias{cov_noise}
\title{Compute a white noise covariance matrix}
\usage{
cov_noise(sigma = 0.5, n)
}
\arguments{
\item{sigma}{\code{numeric} scalar denoting the variance. Defaults to \code{0.5}}

\item{n}{\code{integer} scalar denoting the number of rows and columns required in the covariance matrix}
}
\value{
\code{GPCov} containing the covariance matrix
}
\description{
Compute a white noise covariance matrix
}
\examples{
cov_noise(0.05, 100)

}
\author{
Trent Henderson
}
