% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_noise_cpp}
\alias{cov_noise_cpp}
\title{Compute a white noise covariance matrix}
\usage{
cov_noise_cpp(sigma, n)
}
\arguments{
\item{sigma}{\code{double} denoting the variance}

\item{n}{\code{int} denoting the number of rows and columns required in the covariance matrix}
}
\value{
\code{NumericMatrix} of covariance
}
\description{
Compute a white noise covariance matrix
}
\examples{
cov_noise_cpp(0.05, 100)
}
\author{
Trent Henderson
}
