% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP.R
\name{GP}
\alias{GP}
\title{Calculate posterior mean vector and covariance matrix}
\usage{
GP(x, xprime, y, covfun, sigma_noise = 0.5, ...)
}
\arguments{
\item{x}{\code{numeric} vector of input data}

\item{xprime}{\code{numeric} vector of data points to predict}

\item{y}{\code{numeric} vector values to learn from}

\item{covfun}{\code{function} specifying the covariance function to use}

\item{sigma_noise}{\code{numeric} scalar denoting the variance of the noise to be added to the covariance matrix. Defaults to \code{0.5}}
}
\value{
\code{TSGP} object containing the input data, posterior mean function and covariance matrix
}
\description{
Calculate posterior mean vector and covariance matrix
}
\examples{
x1 <- 1:100
y <- 3 * sin(2 * seq(0, 4 * pi, length.out = 100)) + runif(100) * 2 + (0.08 * seq(from = 1, to = 100, by = 1))

CovSum <- function(xa, xb, sigma_1 = 1, sigma_2 = 1, l_1 = 1, l_2 = 1, p = 1){
  Sigma_exp_quad <- cov_exp_quad(xa, xb, sigma_1, l_1)
  Sigma_periodic <- cov_periodic(xa, xb, sigma_2, l_2, p)
  Sigma <- Sigma_exp_quad + Sigma_periodic
  return(Sigma)
}

mod <- GP(x1, 1:length(y), y, CovSum, 0.5)

}
\author{
Trent Henderson
}
